Improved based off feedback
- split classes: SwipeListener and SwipeActivity
- changed gestureDetector to take in context
- SwiperListener(this, relativeLayout)

Currently crashing. Not sure where I went wrong in splitting the classes
